campaign-planner: Break down planner data in 4 chunks to prevent server memory overflow
Campaign Planner data are now accessible through 4 routes.
This applies for both logged in users and guests

For guest users, the campaign planner save is now available separately from the planner data, but a valid save uid is required to get access to the planner data
Guest rate limit has been bumped from 10 to 60 as the number of requests to open the planner as gone from 1 to 5.

#CONNECT-165 State In Progress