broadcast: Promote Schedule script improvement;
Add an retry mechanism for schedules promotion in case no ID can be found in the campaign. This is because some schedule promotion runs before the campaign has has the time to be promoted, making them end with failure.
With this new mechanism, a schedule promotion without any specific representation given, failing because the campaign has no external resource matching a representation, will register a new PromoteScheduleJob specific to the current representation to try again later.